backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Werkzeug
pandas
scikit-learn 
joblib